refactor: move download file impl to a separate file
It grew quite complicated, well worth it to isolate, especially given
that the public API is tiny -- just a single function.
